Profile Details Full

+++ lib/features/enrolled_courses/model/course_detail_model.dart
@@ -39,7 +39,7 @@ class Data {
   List<String>? likes;
   int? likeCount;
   int? ratingCount;
-  int? averageRating;
+  double? averageRating;
 
   Data(
       {this.id,
